[cleanup] moved `CRRequest` into its own file
